adding new metrics
