+++ huginn-proxy-lib/src/proxy/proxy_protocol.rs
@@ -1,12 +1,18 @@
-//! Minimal PROXY protocol **v2** reader.
+//! Source address recovery for L4-forwarded connections.
 //!
-//! Recovers the original client `(src_ip, src_port)` when huginn runs behind an L4 proxy that
-//! does TLS passthrough (e.g. Traefik `IngressRouteTCP` with `proxyProtocol.version: 2`). The
-//! parsed source address replaces the socket peer for the eBPF SYN lookup, `X-Forwarded-*`,
-//! rate-limiting, IP filtering and logs. See `data/proxy-protocol.md` for the design.
+//! When huginn sits behind any PROXY-protocol-capable load balancer (HAProxy, nginx stream,
+//! AWS Network Load Balancer, Envoy, Kubernetes L4 ingress) the TCP socket peer is the
+//! load balancer, not the client. This module parses the forwarding header prepended by the
+//! load balancer to recover the original client `(src_ip, src_port)`.
 //!
-//! v1 (text) is intentionally **not** supported: v2 is a fixed binary layout, so we can consume
-//! **exactly** the header length and leave the stream aligned on the following TLS ClientHello.
+//! The recovered address replaces the socket peer for eBPF SYN correlation, `X-Forwarded-*`,
+//! rate-limiting, IP filtering and logs. See `data/proxy-protocol.md` for the full design.
+//!
+//! Only the **binary encoding** is supported. The text encoding is intentionally omitted:
+//! the binary signature (`\r\n\r\n\0\r\nQUIT\n`) cannot collide with a TLS ClientHello
+//! (`0x16 0x03…`) or HTTP, enabling non-destructive detection via `MSG_PEEK`. The declared
+//! address-block length lets the parser consume **exactly** those bytes, leaving the stream
+//! aligned on the following TLS ClientHello.
 
 use std::fmt;
 use std::net::{Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6};
@@ -172,144 +178,5 @@ pub async fn looks_like_proxy_v2(stream: &TcpStream) -> std::io::Result<bool> {
         }
         // A trusted peer mid-send: yield and re-peek (bounded by the caller's timeout).
         tokio::task::yield_now().await;
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io::Cursor;
-
-    type TestResult = Result<(), Box<dyn std::error::Error>>;
-
-    /// Build a v2 header: signature + ver/cmd + fam/proto + len + address block.
-    fn build_header(cmd: u8, fam: u8, addrs: &[u8]) -> Vec<u8> {
-        let mut buf = Vec::new();
-        buf.extend_from_slice(&V2_SIGNATURE);
-        buf.push((2 << 4) | (cmd & 0x0F));
-        buf.push((fam << 4) | 0x1); // protocol = STREAM (0x1)
-        buf.extend_from_slice(&(addrs.len() as u16).to_be_bytes());
-        buf.extend_from_slice(addrs);
-        buf
-    }
-
-    fn v4_block(src: [u8; 4], src_port: u16) -> Vec<u8> {
-        let mut b = Vec::new();
-        b.extend_from_slice(&src); // src
-        b.extend_from_slice(&[10, 0, 0, 1]); // dst (ignored)
-        b.extend_from_slice(&src_port.to_be_bytes()); // src port
-        b.extend_from_slice(&8443u16.to_be_bytes()); // dst port (ignored)
-        b
-    }
-
-    fn v6_block(src: [u8; 16], src_port: u16) -> Vec<u8> {
-        let mut b = Vec::new();
-        b.extend_from_slice(&src); // src
-        b.extend_from_slice(&[0u8; 16]); // dst (ignored)
-        b.extend_from_slice(&src_port.to_be_bytes()); // src port
-        b.extend_from_slice(&8443u16.to_be_bytes()); // dst port (ignored)
-        b
-    }
-
-    #[tokio::test]
-    async fn parses_valid_ipv4() -> TestResult {
-        let header = build_header(CMD_PROXY, FAM_INET, &v4_block([192, 168, 1, 100], 51234));
-        let mut cursor = Cursor::new(header);
-        let src = read_proxy_header_v2(&mut cursor).await?;
-        assert_eq!(src, Some("192.168.1.100:51234".parse()?));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn parses_valid_ipv6() -> TestResult {
-        let octets = Ipv6Addr::new(0x2001, 0xdb8, 0, 0, 0, 0, 0, 1).octets();
-        let header = build_header(CMD_PROXY, FAM_INET6, &v6_block(octets, 40000));
-        let mut cursor = Cursor::new(header);
-        let src = read_proxy_header_v2(&mut cursor).await?;
-        assert_eq!(src, Some("[2001:db8::1]:40000".parse()?));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn local_command_returns_none() -> TestResult {
-        // LOCAL command (health checks): no address block.
-        let header = build_header(CMD_LOCAL, 0x0, &[]);
-        let mut cursor = Cursor::new(header);
-        let src = read_proxy_header_v2(&mut cursor).await?;
-        assert_eq!(src, None);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn af_unspec_falls_back() -> TestResult {
-        // PROXY command but AF_UNSPEC family → no usable source.
-        let header = build_header(CMD_PROXY, 0x0, &[]);
-        let mut cursor = Cursor::new(header);
-        let src = read_proxy_header_v2(&mut cursor).await?;
-        assert_eq!(src, None);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn bad_signature_is_rejected() {
-        let mut bytes = vec![0u8; FIXED_LEN];
-        bytes[0] = 0xFF; // corrupt the signature
-        let mut cursor = Cursor::new(bytes);
-        let result = read_proxy_header_v2(&mut cursor).await;
-        assert!(matches!(result, Err(ProxyProtocolError::BadSignature)), "got {result:?}");
-    }
-
-    #[tokio::test]
-    async fn v1_text_header_is_unsupported_version() {
-        // A v1 text header ("PROXY TCP4 ...") does not match the binary signature.
-        let mut cursor = Cursor::new(b"PROXY TCP4 1.2.3.4 5.6.7.8 1 2\r\n".to_vec());
-        let result = read_proxy_header_v2(&mut cursor).await;
-        assert!(matches!(result, Err(ProxyProtocolError::BadSignature)), "got {result:?}");
-    }
-
-    #[tokio::test]
-    async fn truncated_address_block_errors() {
-        // Announce AF_INET but provide fewer than 12 address bytes.
-        let header = build_header(CMD_PROXY, FAM_INET, &[1, 2, 3, 4]);
-        let mut cursor = Cursor::new(header);
-        let result = read_proxy_header_v2(&mut cursor).await;
-        assert!(matches!(result, Err(ProxyProtocolError::Truncated)), "got {result:?}");
-    }
-
-    #[tokio::test]
-    async fn consumes_exactly_header_leaving_clienthello() -> TestResult {
-        // Header followed by a synthetic TLS ClientHello prefix. After parsing, the next bytes
-        // read must equal the ClientHello untouched (alignment proof).
-        let client_hello = [0x16u8, 0x03, 0x01, 0x00, 0x2a, 0xde, 0xad, 0xbe, 0xef];
-        let mut stream = build_header(CMD_PROXY, FAM_INET, &v4_block([203, 0, 113, 5], 12345));
-        stream.extend_from_slice(&client_hello);
-
-        let mut cursor = Cursor::new(stream);
-        let src = read_proxy_header_v2(&mut cursor).await?;
-        assert_eq!(src, Some("203.0.113.5:12345".parse()?));
-
-        let mut rest = Vec::new();
-        cursor.read_to_end(&mut rest).await?;
-        assert_eq!(rest, client_hello, "ClientHello bytes must remain after the header");
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn ignores_trailing_tlvs_but_consumes_them() -> TestResult {
-        // Address block longer than the minimum (TLVs appended) must be fully consumed.
-        let mut block = v4_block([10, 1, 2, 3], 5555);
-        block.extend_from_slice(&[0x03, 0x00, 0x02, 0xAA, 0xBB]); // a fake TLV
-        let client_hello = [0x16u8, 0x03, 0x03];
-        let mut stream = build_header(CMD_PROXY, FAM_INET, &block);
-        stream.extend_from_slice(&client_hello);
-
-        let mut cursor = Cursor::new(stream);
-        let src = read_proxy_header_v2(&mut cursor).await?;
-        assert_eq!(src, Some("10.1.2.3:5555".parse()?));
-
-        let mut rest = Vec::new();
-        cursor.read_to_end(&mut rest).await?;
-        assert_eq!(rest, client_hello);
-        Ok(())
     }
 }
